fix dto

+++ app/src/main/java/com/elmorshdi/weatheraplication/data/remote/model/Forecast.kt
@@ -1,11 +1,13 @@
 package com.elmorshdi.weatheraplication.data.remote.model
 
+
+
 data class Forecast (
     val clouds: Clouds?,
     val dt: Int?,
     val dt_txt: String?,
     val main: Main?,
-    val pop: Int?,
+    val pop: Double?,
     val sys: Sys?,
     val visibility: Int?,
     val weather: List<Weather?>?,
